Move upload_file and clear_metadata to storage.py
