+++ app/views/subs.py
@@ -12,7 +12,7 @@ from ..socketio import socketio
 from ..models import Sub, db as pdb, SubMod, SubMetadata, SubStylesheet, SubSubscriber, SiteMetadata, SubPost
 from ..models import SubPostPollOption, SubPostMetadata, SubPostVote, User, UserUploads
 from ..forms import CreateSubPostForm, CreateSubForm
-from ..storage import file_url
+from ..storage import file_url, upload_file
 
 bp = Blueprint('subs', __name__)
 
@@ -152,7 +152,7 @@ def create_post(ptype, sub):
     if form.ptype.data in ('link', 'upload'):
         # TODO: Make a different ptype for uploads?
         ptype = 1
-        fupload = misc.upload_file()
+        fupload = upload_file()
         if fupload[0] is not False and fupload[1] is False:
             return engine.get_template('sub/createpost.html').render(
                 {'error': fupload[0], 'form': form, 'sub': sub, 'captcha': captcha}), 400
